fix: clean up agent refactoring and test updates
