+++ test_agent_ollama.py
@@ -8,7 +8,7 @@ Requirements: Ollama running with 'mistral' model
 
 import asyncio
 import json
-from ai.agent import Agent, AgentConfig
+from ai.agent_ollama import Agent, AgentConfig
 from ai.rules import PublicState
